Protect server with rate limit
Set rate limit on 30 (per minute per IP). This allows our users that want to
export a single or few charts (using the hamburger menu) to do so 30 times a
minute, which should be enough in normal situations. It protects our server from
DoS attacks by one or a few servers though.

We also set a skipKey and skipToken which we can supply to bypass the rateLimit
when calling the export server from code, e.g. in migrations or when converting
the charts in a report. When the report is very large (with more than 30
charts) a single refresh of the whole report could trigger the rate limit
otherwise.

The key/token can be provided in the URL with `?key=smarttrackers&access_token=<token>`

Normally we shouldn't include these tokens in source code tracked by Git, but
the security risk for this project is so minimal we don't really care in this case.